[DEV-8580] add text width adjustments for sankey (#1673)
* [DEV-9710] add margin on existing dropdown filter button

* [DEV-8580] remove unused vars

* [DEV-8580] updating types and vars

* [DEV-8580] update tooltip helper

* [DEV-8580] minor refactors

* [DEV-8580] more refactors

* [DEV-8580] Remove need for message

* [DEV-8580] update font sizes
