+++ packages/chart/index.html
@@ -26,11 +26,11 @@
 
   <body>
     <!-- SANKEY EXAMPLE -->
-    <!-- <div class="container d-flex flex-wrap body-wrapper bg-white">
+    <!-- <div class="container d-flex flex-wrap body-wrapper bg-white" style="width:  750px">
       <main class="col-lg-9 order-lg-2" role="main" aria-label="Main Content Area">
         <div class="row">
           <div class="col-md-12"> -->
-    <!-- <div class="react-container" data-config="/examples/feature/sankey/sankey-example-data.json"></div> -->
+    <div class="react-container" data-config="/examples/feature/sankey/sankey-example-data.json"></div>
     <!-- </div>
         </div>
       </main>
@@ -57,7 +57,6 @@
     <!-- <div class="react-container" data-config="/examples/feature/filters/url-filter.json"></div> -->
     <!-- <div class="react-container" data-config="/examples/feature/bar/additional-column-tooltip.json"></div> -->
     <!-- <div class="react-container" data-config="https://cdc.gov/poxvirus/mpox/modules/data-viz/mpx-trends_1.json"></div> -->
-    <!-- <div class="react-container" data-config="/examples/feature/area/area-chart-date-city-temperature.json"></div> -->
     <!-- <div class="react-container" data-config="/examples/feature/area/area-chart-date-apple.json"></div> -->
     <!-- <div class="react-container" data-config="/examples/feature/forest-plot/linear.json"></div> -->
     <!-- <div class="react-container" data-config="/examples/feature/forest-plot/logarithmic.json"></div> -->
@@ -74,7 +73,7 @@
     <!-- <div class="react-container" data-config="/examples/feature/area/area-chart-category.json"></div> -->
     <!-- <div class="react-container" data-config="/examples/feature/scatterplot/scatterplot.json"></div> -->
     <!-- <div class="react-container" data-config="/examples/feature/deviation/planet-deviation-config.json"></div> -->
-    <div class="react-container" data-config="/examples/feature/boxplot/boxplot.json"></div>
+    <!-- <div class="react-container" data-config="/examples/feature/boxplot/boxplot.json"></div> -->
     <!-- <div class="react-container" data-config="/examples/feature/combo/planet-combo-example-config.json"></div> -->
     <!-- <div class="react-container" data-config="/examples/feature/combo/right-issues.json"></div> -->
     <!-- <div class="react-container" data-config="/examples/region-issue.json"></div> -->
@@ -105,7 +104,10 @@
     <!-- <div class="react-container" data-config="/https://www.cdc.gov/poxvirus/mpox/modules/data-viz/mpx-age-sex1.json"></div> -->
 
     <!-- TESTS DATE EXCLUSIONS -->
-    <!-- <div class="react-container" data-config="/examples/feature/tests-date-exclusions/date-exclusions-config.json"></div> -->
+    <div
+      class="react-container"
+      data-config="/examples/feature/tests-date-exclusions/date-exclusions-config.json"
+    ></div>
     <!-- <div class="react-container" data-config="/examples/feature/tests-case-rate/case-rate-example-config.json"></div> -->
 
     <!-- TESTS BIG SMALL-->
